src/main/java/de/unipotsdam/cs/toolup/ws/resource/ApplicationResource.java
package de.unipotsdam.cs.toolup.ws.resource;

import de.unipotsdam.cs.toolup.ws.beans.ApplicationBean;

import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.core.MediaType;
import javax.ws.rs.core.Response;

import static de.unipotsdam.cs.toolup.util.JerseyUtil.createResponseOk;


@Path("application")
public class ApplicationResource {

    @GET
    @Path("/")
    @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
    public Response getAll() throws Exception {
        return createResponseOk(ApplicationBean.getAllApplications());
    }

    @GET
    @Path("/{id}")
    @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
    public Response get(@PathParam("id") String id) throws Exception {
        return createResponseOk(ApplicationBean.getBean(id));
    }

}
